backend/rapidtires/requirements.txt
django>=4.2,<5.2
djangorestframework
django-cors-headers
whitenoise
gunicorn
python-decouple
psycopg2-binary
